Added protocol 1.4.6 documentation (CashToken support)
Document the API changes in 1.4.6 (upcoming Fulcrum 1.9.0). This
includes addition of `token_data` to some of the results e.g. to
listunspent, as well as other changes.

Also removed the "protocol-ideas.rst" file since it was very much
outdated.

Also deleted some of the methods in `protocol-removed.rst` that we have
since added back to the protocol (such as the `blockchain.address.*`
methods).