Correct non-verbose blockchain.transaction.get() output
The output includes quotes around the raw transaction in a single line.  While wrapping might be desirable from a display perspective, there aren't quotes around multiple lines in the output.  I changed it to a single line because manually adding line breaks means the copy function won't return an actual raw transaction.  Based on a cursory search, there are numerous ways to make it wrap, but I'm not sure which would apply or work well in this case, so I'm hesitant to randomly choose one.